% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macro_tx2gene}
\alias{macro_tx2gene}
\title{Macrophage dataset - transcript-to-gene table}
\format{
A \code{DataFrame} with 205870 rows and columns
\code{tx_name} and \code{gene_id}.
}
\source{
Gencode v29 human reference transcripts
}
\usage{
data("macro_tx2gene")
}
\description{
A table mapping all transcripts of the Gencode v29 human reference
(as indexed by Salmon for the macrophage dataset) to genes. It can be
used to summarize \code{\link{macro_txp_se}} to the gene level with
\code{tximeta::summarizeToGene(se, skipRanges=TRUE, tx2gene=macro_tx2gene)},
which does not require the GTF file. For the script used to build this
object, see \code{makeSumExp.R} in the \code{scripts} directory of the
installed package.
}
\seealso{
\code{\link{macro_txp_se}}
}
\keyword{datasets}
