\docType{package}
\name{macrophage-package}
\alias{macrophage-package}
\title{Salmon quantifications for human macrophage immune response}
\description{
  This package provides the output of running Salmon on a set
  of 24 RNA-seq samples from Alasoo, et al. (2018).
  For more details on the data and steps used to generate the
  quantification files, please refer to the package vignette.
}
\references{
  Alasoo, et al. "Shared genetic effects on chromatin and gene
  expression indicate a role for enhancer priming in immune response",
  Nature Genetics, January 2018 doi: 10.1038/s41588-018-0046-7.
}
\keyword{package}
